+++ test/support/serializer.ex
@@ -1,0 +1,6 @@
+defmodule GuardianDb.Test.Serializer do
+  @behaviour Guardian.Serializer
+
+  def for_token(sub), do: {:ok, sub}
+  def from_token(sub), do: {:ok, sub}
+end
